fix: correct elvish shell syntax for `Elvish.format_path_export()`
`set` is a special command for assigning values to variables in Elvish.
It requires an `=` to separate the variable name from the value.

The use of `=` is necessary because it allows for multiple assignments,
such as `set x y z = 1 2 3`, in my opinion.

For more details, refer to the Elvish documentation on [`set`](https://elv.sh/ref/language.html#set).

`$paths` is a variable containing a list of strings that are always synchronized with the `PATH` environment variable. Since it is a variable, we use `set` to assign values to it. Therefore, the correct way to assign values to `$paths` is:

```elvish
set paths = [$E:PROTO_HOME/shims $E:PROTO_HOME/bin $@paths]
```

For more details, refer to the Elvish documentation on [`$paths`](https://elv.sh/ref/builtin.html#$paths).